MAJOR BREAKING UPDATE: Thea now uses Eigen for all linear algebra. Many APIs are now different (not always for the better, e.g. swizzle operators are gone and template argument deduction often requires more casting, eval() and finished()), and everything builds more slowly. On the plus side we can now take advantage of Eigen's consistency, ubiquity and reliability, use the array of solvers that come bundled with Eigen, and the syntax isn't dramatically different. Another major advantage is that fixed and dynamic size matrices can be freely mixed. Along with this change, many interfaces have been reworked -- e.g. nearly all the plugin interfaces (except RenderSystem) now pass only POD types or pure virtual wrappers to allow plugins from different compilers. This update WILL BREAK YOUR CODE, so consider the previous version the 'frozen' version of the old API.

+++ Code/Source/Algorithms/IntersectionTester_Transform.hpp
@@ -49,7 +49,7 @@ namespace Thea {
 namespace Algorithms {
 
 // Default specializations
-template <typename ObjectA, typename TransformA, typename ObjectB, typename TransformB, long N, typename T>
+template <typename ObjectA, typename TransformA, typename ObjectB, typename TransformB, int N, typename T>
 struct IntersectionTesterImpl< TransformedObject<ObjectA, TransformA>, TransformedObject<ObjectB, TransformB>, N, T >
 {
   typedef TransformedObject<ObjectA, TransformA> TA;
@@ -77,7 +77,7 @@ struct IntersectionTesterImpl< TransformedObject<ObjectA, TransformA>, Transform
   }
 };
 
-template <typename ObjectA, typename TransformA, typename B, long N, typename T>
+template <typename ObjectA, typename TransformA, typename B, int N, typename T>
 struct IntersectionTesterImpl< TransformedObject<ObjectA, TransformA>, B, N, T >
 {
   typedef TransformedObject<ObjectA, TransformA> TA;
@@ -91,7 +91,7 @@ struct IntersectionTesterImpl< TransformedObject<ObjectA, TransformA>, B, N, T >
   }
 };
 
-template <typename A, typename ObjectB, typename TransformB, long N, typename T>
+template <typename A, typename ObjectB, typename TransformB, int N, typename T>
 struct IntersectionTesterImpl< A, TransformedObject<ObjectB, TransformB>, N, T >
 {
   typedef TransformedObject<ObjectB, TransformB> TB;
